# openapi.yaml
openapi: 3.1.0
info:
  title: Scheduling App API
  version: 1.0.0
  description: |
    REST API for the Scheduling App — helps American Airlines flight attendants
    build optimized bidding schedules from their monthly bid sheet, enforcing
    rules from the 2024 AA/APFA Collective Bargaining Agreement (CBA).

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a valid
    Bearer token in the `Authorization` header.

    ## Error Envelope
    All error responses use a standard JSON body:
    ```json
    { "code": "string", "message": "string", "details": {} }
    ```

    ## Pagination
    List endpoints return `{ data: [...], page_state: "token|null" }`.
    Pass `page_state` as a query parameter to fetch the next page.

    ## Requirement Traceability
    Each operation includes an `x-req-ids` extension listing the requirement
    IDs from requirements.md that it fulfills.
servers:
  - url: https://api.schedulingapp.example.com/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [Auth]
      summary: Register a new user
      x-req-ids: [REQ-008, REQ-025]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: loginUser
      tags: [Auth]
      summary: Authenticate and obtain tokens
      x-req-ids: [REQ-008, REQ-025]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Users ───────────────────────────────────────────────────────────────────
...
  /bid-periods/{bidPeriodId}:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    get:
      operationId: getBidPeriod
      tags: [Bid Periods]
      summary: Get a single bid period with parsing status
      x-req-ids: [REQ-019, REQ-022]
      responses:
        "200":
          description: Bid period details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidPeriod"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteBidPeriod
      tags: [Bid Periods]
      summary: Delete a bid period and all associated data
      x-req-ids: [REQ-019]
      responses:
        "204":
          description: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /bid-periods/{bidPeriodId}/awarded-schedule:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    post:
      operationId: importAwardedSchedule
      tags: [Awarded Schedule]
      summary: Import the awarded schedule for this bid period
      x-req-ids: [REQ-017]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Awarded schedule document
                bid_id:
                  type: string
                  format: uuid
                  description: The bid to compare against (optional)
      responses:
        "201":
          description: Awarded schedule imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwardedSchedule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    get:
      operationId: getAwardedSchedule
      tags: [Awarded Schedule]
      summary: Get the imported awarded schedule
      x-req-ids: [REQ-017]
      responses:
        "200":
          description: Awarded schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwardedSchedule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /bid-periods/{bidPeriodId}/award-analysis:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    get:
      operationId: getAwardAnalysis
      tags: [Awarded Schedule]
      summary: Analyze bid-vs-award performance
      x-req-ids: [REQ-018]
      parameters:
        - name: bid_id
          in: query
          schema:
            type: string
            format: uuid
          description: Specific bid to compare against (defaults to the finalized bid)
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwardAnalysis"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

# ═══════════════════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BidPeriodId:
      name: bidPeriodId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists (e.g. duplicate email, duplicate bookmark)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: Request understood but cannot be processed (e.g. unparseable PDF)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: ["VALIDATION_ERROR", "NOT_FOUND", "PARSE_FAILED"]
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          additionalProperties: true
          description: Additional context (field errors, etc.)

    # ── Auth ───────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, profile]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        profile:
          $ref: "#/components/schemas/ProfileInput"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Seconds until access_token expires
        user:
          $ref: "#/components/schemas/User"

    # ── User ───────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        profile:
          $ref: "#/components/schemas/Profile"
        default_preferences:
          $ref: "#/components/schemas/Preferences"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        display_name:
          type: string
        base_city:
          type: string
          description: IATA code
        seniority_number:
          type: integer
          description: "Lower = more senior"
        total_base_fas:
          type: integer
          description: "Total FAs at this base; used with seniority_number to compute percentile"
        position_min:
          type: integer
        position_max:
          type: integer
        language_qualifications:
          type: array
          items:
            type: string
          description: "e.g. ['JP', 'SP']"
        commute_from:
          type: string
          nullable: true
          description: "IATA code of city FA commutes from (e.g. 'DCA'). Used for commute impact analysis on sequences."
        years_of_service:
          type: integer
          description: "Years of service for CBA §3.A pay rate lookup (1-13+)"
        is_reserve:
          type: boolean
          description: "true = Reserve, false = Lineholder"
        is_purser_qualified:
          type: boolean
          description: "Purser qualification per CBA §14.L"
        line_option:
          type: string
          enum: [standard, high, low]
          description: "Line of Time option: standard (70-90h), high (up to 110h), low (down to 40h) per CBA §2.EE"
    ProfileInput:
      type: object
      required: [display_name, base_city, seniority_number, total_base_fas, position_min, position_max]
      properties:
        display_name:
          type: string
        base_city:
          type: string
        seniority_number:
          type: integer
        total_base_fas:
          type: integer
        position_min:
          type: integer
        position_max:
          type: integer
        language_qualifications:
          type: array
          items:
            type: string
        commute_from:
          type: string
          nullable: true
          description: "IATA code of city FA commutes from (e.g. 'DCA')"
        years_of_service:
          type: integer
        is_reserve:
          type: boolean
          default: false
        is_purser_qualified:
          type: boolean
          default: false
        line_option:
          type: string
          enum: [standard, high, low]
          default: standard
...
    Preferences:
      type: object
      description: All fields are optional; omitted fields retain their current value.
      properties:
        preferred_days_off:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 31
        preferred_layover_cities:
          type: array
          items:
            type: string
        avoided_layover_cities:
          type: array
          items:
            type: string
        tpay_min_minutes:
          type: integer
        tpay_max_minutes:
          type: integer
        preferred_equipment:
          type: array
          items:
            type: string
        report_earliest_minutes:
          type: integer
          description: Minutes from midnight
        report_latest_minutes:
          type: integer
        release_earliest_minutes:
          type: integer
        release_latest_minutes:
          type: integer
        avoid_redeyes:
          type: boolean
        prefer_turns:
          type: boolean
          nullable: true
        prefer_high_ops:
          type: boolean
          nullable: true
        cluster_trips:
          type: boolean
        prefer_holidays:
          type: boolean
          nullable: true
          description: "Prefer sequences operating on CBA §3.K holidays (100% premium)"
        prefer_speaker_sequences:
          type: boolean
          nullable: true
          description: "For language-qualified FAs: prefer Speaker sequences"
        weights:
          $ref: "#/components/schemas/PreferenceWeights"
...
    BidPeriod:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        effective_start:
          type: string
          format: date
        effective_end:
          type: string
          format: date
        base_city:
          type: string
        source_filename:
          type: string
        parse_status:
          type: string
          enum: [pending, processing, completed, failed]
        parse_error:
          type: string
          nullable: true
        total_sequences:
          type: integer
        categories:
          type: array
          items:
            type: string
        issued_date:
          type: string
          format: date
          nullable: true
        preference_overrides:
          $ref: "#/components/schemas/Preferences"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AwardedSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bid_period_id:
          type: string
          format: uuid
        bid_id:
          type: string
          format: uuid
          nullable: true
        source_filename:
          type: string
        imported_at:
          type: string
          format: date-time
        awarded_sequences:
          type: array
          items:
            $ref: "#/components/schemas/AwardedSequenceEntry"
    AwardedSequenceEntry:
      type: object
      properties:
        seq_number:
          type: integer
        sequence_id:
          type: string
          format: uuid
          nullable: true
        operating_dates:
          type: array
          items:
            type: integer
        tpay_minutes:
          type: integer
        block_minutes:
          type: integer
        tafb_minutes:
          type: integer
    AwardAnalysis:
      type: object
      properties:
        bid_id:
          type: string
          format: uuid
        awarded_schedule_id:
          type: string
          format: uuid
        match_count:
          type: integer
          description: Sequences in the bid that were awarded
        match_rate:
          type: number
          format: float
          description: "0.0 to 1.0"
        top_10_match_count:
          type: integer
        top_10_match_rate:
          type: number
          format: float
        matched_entries:
          type: array
          items:
            type: object
            properties:
              seq_number:
                type: integer
              bid_rank:
                type: integer
              was_awarded:
                type: boolean
              attainability:
                type: string
                enum: [high, medium, low, unknown]
                description: "What the optimizer predicted for this sequence"
        unmatched_awards:
          type: array
          items:
            type: integer
          description: SEQ numbers awarded but not in the bid
        attainability_accuracy:
          type: object
          description: "How well the optimizer's seniority estimates performed"
          properties:
            high_awarded:
              type: integer
              description: "Sequences marked 'high' attainability that were actually awarded"
            high_total:
              type: integer
            low_awarded:
              type: integer
              description: "Sequences marked 'low' that were awarded (surprises)"
            low_total:
              type: integer
        insights:
          type: array
          items:
            type: string
          description: Human-readable insights about bid performance and seniority calibration
...